[clean] divide files
