+++ main.c
@@ -1,0 +1,17 @@
+#include "slcc.h"
+
+int main(int argc, char **argv) {
+  if (argc != 2) {
+    fprintf(stderr, "引数の個数が正しくありません\n");
+    return 1;
+  }
+
+  // トークナイズする
+  user_input = argv[1];
+  token = tokenize(user_input);
+  Node *node = expr();
+
+  codegen(node);
+
+  return 0;
+}
