--- slcc.c
@@ -1,377 +1,0 @@
-#include <ctype.h>
-#include <stdarg.h>
-#include <stdbool.h>
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-
-// 抽象構文木のノードの種類
-typedef enum {
-  ND_ADD, // +
-  ND_SUB, // -
-  ND_MUL, // *
-  ND_DIV, // /
-  ND_EQ,  // ==
-  ND_NE,  // !=
-  ND_LT,  // <
-  ND_LE,  // <=
-  ND_NUM, // 0-9
-} NodeKind;
-
-typedef struct Node Node;
-
-// 抽象構文木のノードの型
-struct Node {
-  NodeKind kind;
-  Node *lhs;
-  Node *rhs;
-  int val;  // ND_NUM のときに使われる
-};
-
-bool consume(char *op);
-Node *primary();
-Node *mul();
-Node *unary();
-Node *expr();
-Node *equality();
-Node *relational();
-Node *add();
-void expect(char *op);
-int expect_number();
-
-
-Node *new_node(NodeKind kind, Node *lhs, Node *rhs) {
-  Node *node = calloc(1, sizeof(Node));
-  node->kind = kind;
-  node->lhs = lhs;
-  node->rhs = rhs;
-  return node;
-}
-
-Node *new_node_num(int val) {
-  Node *node = calloc(1, sizeof(Node));
-  node->kind = ND_NUM;
-  node->val = val;
-  return node;
-}
-
-// expr = equality
-Node *expr() {
-  return equality();
-}
-
-// equality = relational ("==" relational | "!=" relational)*
-Node *equality() {
-  Node *node = relational();
-
-  for (;;) {
-    if (consume("==")) {
-      node = new_node(ND_EQ, node, relational());
-    }
-    else if (consume("!=")) {
-      node = new_node(ND_NE, node, relational());
-    }
-    else {
-      return node;
-    }
-  }
-}
-
-// relational = add ("<" add | "<=" add | ">" add | ">=" add)*
-Node *relational() {
-  Node *node = add();
-
-  for (;;) {
-    if (consume("<")) {
-      node = new_node(ND_LT, node, add());
-    }
-    else if (consume("<=")) {
-      node = new_node(ND_LE, node, add());
-    }
-    else if (consume(">")) {
-      node = new_node(ND_LT, add(), node);
-    }
-    else if (consume(">=")) {
-      node = new_node(ND_LE, add(), node);
-    }
-    else {
-      return node;
-    }
-  }
-}
-
-// add = mul ("+" mul | "-" mul)*
-Node *add() {
-  Node *node = mul();
-
-  for (;;) {
-    if (consume("+")) {
-      node = new_node(ND_ADD, node, mul());
-    }
-    else if (consume("-")) {
-      node = new_node(ND_SUB, node, mul());
-    }
-    else {
-      return node;
-    }
-  }
-}
-
-// mul = unary ( "*" primary | "/" primary )*
-Node *mul() {
-  Node *node = unary();
-
-  for (;;) {
-    if (consume("*")) {
-      node = new_node(ND_MUL, node, primary());
-    }
-    else if (consume("/")) {
-      node = new_node(ND_DIV, node, primary());
-    }
-    else {
-      return node;
-    }
-  }
-}
-
-// unary = ("+" | "-")? primary
-Node *unary() {
-  if (consume("+")) {
-    // +x -> x
-    return primary();
-  }
-  if (consume("-")) {
-    // -x -> 0 - x
-    return new_node(ND_SUB, new_node_num(0), primary());
-  }
-  return primary();
-}
-
-// primary = num | "(" expr ")"
-Node *primary() {
-  // ( が来たら "(" expr ")" が来るはず、来なければ受理できない
-  if (consume("(")) {
-    Node *node = expr();
-    expect(")");
-    return node;
-  }
-
-  // そうでなければ数値を受理
-  return new_node_num(expect_number());
-}
-
-void gen(Node *node) {
-  if (node->kind == ND_NUM) {
-    printf("  push %d\n", node->val);
-    return;
-  }
-
-  gen(node->lhs);
-  gen(node->rhs);
-
-  printf("  pop rdi\n");
-  printf("  pop rax\n");
-
-  switch(node->kind) {
-    case ND_ADD:
-      printf("  add rax, rdi\n");
-      break;
-    case ND_SUB:
-      printf("  sub rax, rdi\n");
-      break;
-    case ND_MUL:
-      printf("  imul rax, rdi\n");
-      break;
-    case ND_DIV:
-      printf("  cqo\n");
-      printf("  idiv rdi\n");
-      break;
-    case ND_EQ:
-      printf("  cmp rax, rdi\n");
-      printf("  sete al\n");
-      printf("  movzb rax, al\n");
-      break;
-    case ND_NE:
-      printf("  cmp rax, rdi\n");
-      printf("  setne al\n");
-      printf("  movzb rax, al\n");
-      break;
-    case ND_LT:
-      printf("  cmp rax, rdi\n");
-      printf("  setl al\n");
-      printf("  movzb rax, al\n");
-      break;
-    case ND_LE:
-      printf("  cmp rax, rdi\n");
-      printf("  setle al\n");
-      printf("  movzb rax, al\n");
-      break;
-  }
-
-  printf("  push rax\n");
-}
-
-// トークンの種類
-typedef enum {
-  TK_RESERVED, // 記号
-  TK_NUM,      // 整数トークン
-  TK_EOF,      // 入力の終わりを表すトークン
-} TokenKind;
-
-typedef struct Token Token;
-
-struct Token {
-  TokenKind kind; // トークンの型
-  Token *next;    // 次の入力トークン
-  int val;        // kindがTK_NUMの場合、その数値
-  char *str;      // トークン文字列
-  int len;        // トークンの長さ
-};
-
-// 入力プログラム
-char *user_input;
-
-// 現在注目しているトークン
-Token *token;
-
-// エラーを報告するための関数
-// printfと同じ引数を取る
-void error(char *fmt, ...) {
-  va_list ap;
-  va_start(ap, fmt);
-  vfprintf(stderr, fmt, ap);
-  fprintf(stderr, "\n");
-  exit(1);
-}
-
-void error_at(char *loc, char *fmt, ...) {
-  va_list ap;
-  va_start(ap, fmt);
-
-  int pos = loc - user_input;
-  fprintf(stderr, "%s\n", user_input);
-  fprintf(stderr, "%*s", pos, " ");
-  fprintf(stderr, "^ ");
-  vfprintf(stderr, fmt, ap);
-  fprintf(stderr, "\n");
-  exit(1);
-}
-
-// 次のトークンが期待している文字列のときには、トークンを1つ読み進めて真を返す。
-// それ以外の場合には偽を返す。
-bool consume(char *op) {
-  if (token->kind != TK_RESERVED ||
-      strlen(op) != token->len ||
-      memcmp(token->str, op, token->len)) {
-    return false;
-  }
-  token = token->next;
-  return true;
-}
-
-// 次のトークンが期待している文字列のときには、トークンを1つ読み進める。
-// それ以外の場合にはエラーを報告する。
-void expect(char *op) {
-  if (token->kind != TK_RESERVED ||
-      strlen(op) != token->len ||
-      memcmp(token->str, op, token->len)) {
-    error_at(token->str, "\"%s\"ではありません", op);
-  }
-  token = token->next;
-}
-
-// 次のトークンが数値の場合、1つ読み進めてその数値を返す。
-// それ以外の場合にはエラーを報告する。
-int expect_number() {
-  if (token->kind!= TK_NUM) {
-    error_at(token->str, "数ではありません");
-  }
-  int val = token->val;
-  token = token->next;
-
-  return val;
-}
-
-bool at_eof() {
-  return token->kind == TK_EOF;
-}
-
-// 新しい Token を cur につなげる
-Token *new_token(TokenKind kind, Token *cur, char *str, int len) {
-  Token *tok = calloc(1, sizeof(Token));
-  tok->kind = kind;
-  tok->str = str;
-  tok->len = len;
-  cur->next = tok;
-
-  return tok;
-}
-
-// p が q で始まるかどうかを判定する
-bool starts_with(char *p, char *q) {
-  return memcmp(p, q, strlen(q)) == 0;
-}
-
-// 入力文字列 user_input をトークナイズして返す
-Token *tokenize(char *p) {
-  Token head;
-  head.next = NULL;
-  Token *cur = &head;
-
-  while (*p) {
-    // 空白文字をスキップ
-    if (isspace(*p)) {
-      p++;
-      continue;
-    }
-
-    // 複数文字の演算子
-    if (starts_with(p, "==") || starts_with(p, "!=") ||
-        starts_with(p, "<=") || starts_with(p, ">=")) {
-      cur = new_token(TK_RESERVED, cur, p, 2);
-      p += 2;
-      continue;
-    }
-
-    if (strchr("+-*/()<>", *p)) {
-      cur = new_token(TK_RESERVED, cur, p++, 1);
-      continue;
-    }
-
-    if (isdigit(*p)) {
-      cur = new_token(TK_NUM, cur, p, 0);
-      char *q = p;
-      cur->val = strtol(p, &p, 10);
-      cur->len = p - q;
-      continue;
-    }
-
-    error_at(p, "不正なトークンです");
-  }
-
-  new_token(TK_EOF, cur, p, 0);
-  return head.next;
-}
-
-int main(int argc, char **argv) {
-  if (argc != 2) {
-    fprintf(stderr, "引数の個数が正しくありません\n");
-    return 1;
-  }
-
-  // トークナイズする
-  user_input = argv[1];
-  token = tokenize(user_input);
-  Node *node = expr();
-
-  // アセンブリの最初の方を出力
-  printf(".intel_syntax noprefix\n");
-  printf(".global main\n");
-  printf("main:\n");
-
-  gen(node);
-
-  printf("  pop rax\n");
-  printf("  ret\n");
-  return 0;
-}
