+++ slcc.h
@@ -1,0 +1,72 @@
+#include <ctype.h>
+#include <stdarg.h>
+#include <stdbool.h>
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+/**
+ * tokenize.c
+*/
+
+// トークンの種類
+typedef enum {
+  TK_RESERVED, // 記号
+  TK_NUM,      // 整数トークン
+  TK_EOF,      // 入力の終わりを表すトークン
+} TokenKind;
+
+typedef struct Token Token;
+struct Token {
+  TokenKind kind; // トークンの型
+  Token *next;    // 次の入力トークン
+  int val;        // kind が TK_NUM の場合、その数値
+  char *str;      // トークン文字列
+  int len;        // トークンの長さ
+};
+
+void error(char *fmt, ...);
+void error_at(char *loc, char *fmt, ...);
+bool consume(char *op);
+void expect(char *op);
+int expect_number();
+bool at_eof();
+Token *new_token(TokenKind kind, Token *cur, char *str, int len);
+Token *tokenize(char *p);
+
+extern char *user_input;
+extern Token *token;
+
+
+/**
+ * parse.c
+*/
+
+// 抽象構文木のノードの種類
+typedef enum {
+  ND_ADD, // +
+  ND_SUB, // -
+  ND_MUL, // *
+  ND_DIV, // /
+  ND_EQ,  // ==
+  ND_NE,  // !=
+  ND_LT,  // <
+  ND_LE,  // <=
+  ND_NUM, // 0-9
+} NodeKind;
+
+// 抽象構文木のノードの型
+typedef struct Node Node;
+struct Node {
+  NodeKind kind;
+  Node *lhs;
+  Node *rhs;
+  int val;  // ND_NUM のときに使われる
+};
+
+Node *expr();
+
+/**
+ * codegen.c
+*/
+void codegen(Node *node);
