+++ tokenize.c
@@ -1,0 +1,130 @@
+#include "slcc.h"
+
+// 入力プログラム
+char *user_input;
+// 現在注目しているトークン
+Token *token;
+
+// エラーを報告するための関数
+// printfと同じ引数を取る
+void error(char *fmt, ...) {
+  va_list ap;
+  va_start(ap, fmt);
+  vfprintf(stderr, fmt, ap);
+  fprintf(stderr, "\n");
+  exit(1);
+}
+
+void error_at(char *loc, char *fmt, ...) {
+  va_list ap;
+  va_start(ap, fmt);
+
+  int pos = loc - user_input;
+  fprintf(stderr, "%s\n", user_input);
+  fprintf(stderr, "%*s", pos, " ");
+  fprintf(stderr, "^ ");
+  vfprintf(stderr, fmt, ap);
+  fprintf(stderr, "\n");
+  exit(1);
+}
+
+// 次のトークンが期待している文字列のときには、トークンを1つ読み進めて真を返す。
+// それ以外の場合には偽を返す。
+bool consume(char *op) {
+  if (token->kind != TK_RESERVED ||
+      strlen(op) != token->len ||
+      memcmp(token->str, op, token->len)) {
+    return false;
+  }
+  token = token->next;
+  return true;
+}
+
+
+// 次のトークンが期待している文字列のときには、トークンを1つ読み進める。
+// それ以外の場合にはエラーを報告する。
+void expect(char *op) {
+  if (token->kind != TK_RESERVED ||
+      strlen(op) != token->len ||
+      memcmp(token->str, op, token->len)) {
+    error_at(token->str, "\"%s\"ではありません", op);
+  }
+  token = token->next;
+}
+
+
+// 次のトークンが数値の場合、1つ読み進めてその数値を返す。
+// それ以外の場合にはエラーを報告する。
+int expect_number() {
+  if (token->kind!= TK_NUM) {
+    error_at(token->str, "数ではありません");
+  }
+  int val = token->val;
+  token = token->next;
+
+  return val;
+}
+
+
+bool at_eof() {
+  return token->kind == TK_EOF;
+}
+
+
+// 新しい Token を cur につなげる
+Token *new_token(TokenKind kind, Token *cur, char *str, int len) {
+  Token *tok = calloc(1, sizeof(Token));
+  tok->kind = kind;
+  tok->str = str;
+  tok->len = len;
+  cur->next = tok;
+
+  return tok;
+}
+
+
+// p が q で始まるかどうかを判定する
+bool starts_with(char *p, char *q) {
+  return memcmp(p, q, strlen(q)) == 0;
+}
+
+// 入力文字列 user_input をトークナイズして返す
+Token *tokenize(char *p) {
+  Token head;
+  head.next = NULL;
+  Token *cur = &head;
+
+  while (*p) {
+    // 空白文字をスキップ
+    if (isspace(*p)) {
+      p++;
+      continue;
+    }
+
+    // 複数文字の演算子
+    if (starts_with(p, "==") || starts_with(p, "!=") ||
+        starts_with(p, "<=") || starts_with(p, ">=")) {
+      cur = new_token(TK_RESERVED, cur, p, 2);
+      p += 2;
+      continue;
+    }
+
+    if (strchr("+-*/()<>", *p)) {
+      cur = new_token(TK_RESERVED, cur, p++, 1);
+      continue;
+    }
+
+    if (isdigit(*p)) {
+      cur = new_token(TK_NUM, cur, p, 0);
+      char *q = p;
+      cur->val = strtol(p, &p, 10);
+      cur->len = p - q;
+      continue;
+    }
+
+    error_at(p, "不正なトークンです");
+  }
+
+  new_token(TK_EOF, cur, p, 0);
+  return head.next;
+}
